+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.ControllerReservationTimeAvailabilityResponse;
-import roomescape.controller.dto.ControllerReservationTimeRequest;
-import roomescape.controller.dto.ControllerReservationTimeResponse;
+import roomescape.controller.dto.request.ControllerReservationTimeRequest;
+import roomescape.controller.dto.response.ControllerReservationTimeAvailabilityResponse;
+import roomescape.controller.dto.response.ControllerReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
-import roomescape.service.dto.ServiceReservationTimeAvailabilityResponse;
-import roomescape.service.dto.ServiceReservationTimeResponse;
+import roomescape.service.dto.response.ServiceReservationTimeAvailabilityResponse;
+import roomescape.service.dto.response.ServiceReservationTimeResponse;
 
 @RestController
 @RequestMapping(value = "/times")
