+++ src/main/java/roomescape/controller/dto/request/ControllerThemeRequest.java
@@ -1,7 +1,7 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
-import roomescape.service.dto.ServiceThemeRequest;
+import roomescape.service.dto.request.ServiceThemeRequest;
 
 public record ControllerThemeRequest(
         @NotBlank(message = "[ERROR] 이름은 비어 있을 수 없습니다.")
