+++ src/main/java/roomescape/controller/dto/response/ControllerReservationTimeResponse.java
@@ -1,7 +1,7 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 import java.time.LocalTime;
-import roomescape.service.dto.ServiceReservationTimeResponse;
+import roomescape.service.dto.response.ServiceReservationTimeResponse;
 
 public record ControllerReservationTimeResponse(Long id, LocalTime startAt) {
     public static ControllerReservationTimeResponse from(ServiceReservationTimeResponse response) {
