+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -10,9 +10,9 @@ import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.dto.ServiceReservationTimeAvailabilityResponse;
-import roomescape.service.dto.ServiceReservationTimeRequest;
-import roomescape.service.dto.ServiceReservationTimeResponse;
+import roomescape.service.dto.request.ServiceReservationTimeRequest;
+import roomescape.service.dto.response.ServiceReservationTimeAvailabilityResponse;
+import roomescape.service.dto.response.ServiceReservationTimeResponse;
 
 @Service
 @Transactional(readOnly = true)
