+++ src/main/java/roomescape/service/dto/request/ServiceReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
